# openapi.yaml
openapi: 3.0.3
info:
  title: Blockvisor API
  description: |
    Blockvisor API is a distributed infrastructure management system that orchestrates software deployment and monitoring across physical and virtual machines. Originally designed for blockchain node management, it has evolved into a general-purpose platform for running and managing any software.
    
    ## Core Concepts
    - **Hosts**: Physical or virtual machines registered with the system via the `bvup` tool
    - **Nodes**: Running software instances with real-time state and metrics monitoring
    - **Organizations**: Groups of users sharing resources and permissions
    - **Users**: Individual accounts with role-based access control (RBAC)
    - **Commands**: Instructions sent to hosts for software lifecycle management
    
    ## Authentication
    The API uses JWT tokens for authentication. Include the token in the `Authorization` header as `Bearer <token>`.
    
    ## Error Handling
    The API returns structured error responses with detailed parameter validation information:
    ```json
    {
      "error": "Invalid query parameters",
      "details": [
        {
          "parameter": "org_id",
          "error": "Invalid UUID format: 'not-a-uuid'",
          "expected": "Valid UUID string (e.g., '550e8400-e29b-41d4-a716-446655440000')"
        }
      ]
    }
    ```
  version: 0.4.0
  contact:
    name: Blockvisor API Support
  license:
    name: MIT
servers:
  - url: https://api.blockvisor.com
...
security:
  - BearerAuth: []
paths:
...
  /v1/protocol:
    get:
      tags:
        - Protocols
      summary: List protocols
      description: Get a list of available protocols
      responses:
        '200':
          description: Protocols retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  protocols:
                    type: array
                    items:
                      $ref: '#/components/schemas/Protocol'

    post:
      tags:
        - Protocols
      summary: Add protocol
      description: Add a new protocol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProtocolRequest'
      responses:
        '201':
          description: Protocol created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  protocol:
                    $ref: '#/components/schemas/Protocol'
  /v1/protocol/{id}:
    get:
      tags:
        - Protocols
      summary: Get protocol
      description: Get details of a specific protocol
      parameters:
        - name: id
          in: path
          required: true
          description: Protocol ID
          schema:
            type: string
            example: "bitcoin"
        - name: org_id
          in: query
          description: Organization ID for private protocols
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Protocol retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  protocol:
                    $ref: '#/components/schemas/Protocol'
        '404':
          description: Protocol not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # API Key endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "An error occurred"
...
    Protocol:
      type: object
      properties:
        id:
          type: string
          example: "bitcoin"
        name:
          type: string
          example: "Bitcoin"
        description:
          type: string
          example: "Bitcoin blockchain protocol"
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
...
    CreateProtocolRequest:
      type: object
      required:
        - key
        - name
      properties:
        key:
          type: string
          example: "bitcoin"
        name:
          type: string
          example: "Bitcoin"
        description:
          type: string
          example: "Bitcoin blockchain protocol"
...